+++ CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 > [!NOTE]
 > While we strive to maintain backwards compatibility as much as possible, we can't guarantee semantic versioning will be strictly followed, as this provider depends on the underlying [bunny.net API](https://docs.bunny.net/reference/bunnynet-api-overview).
 
+## [Unreleased]
+### Fixed
+- resource pullzone_shield: missing realtime_threat_intelligence causes a panic
+
 ## [0.7.3] - 2025-06-17
 ### Added
 - resource pullzone_edgerule: added `priority` to control the execution order of the edge rules ([#37](https://github.com/BunnyWay/terraform-provider-bunnynet/issues/37));